+++ src/query/sql/src/executor/physical_plans/mod.rs
@@ -34,6 +34,7 @@ mod physical_expression_scan;
 mod physical_filter;
 mod physical_hash_join;
 mod physical_join;
+mod physical_join_filter;
 mod physical_limit;
 mod physical_multi_table_insert;
 mod physical_mutation;
@@ -79,9 +80,10 @@ pub use physical_exchange_source::ExchangeSource;
 pub use physical_expression_scan::ExpressionScan;
 pub use physical_filter::Filter;
 pub use physical_hash_join::HashJoin;
-pub use physical_hash_join::PhysicalRuntimeFilter;
-pub use physical_hash_join::PhysicalRuntimeFilters;
 pub use physical_join::PhysicalJoinType;
+pub use physical_join_filter::JoinRuntimeFilter;
+pub use physical_join_filter::PhysicalRuntimeFilter;
+pub use physical_join_filter::PhysicalRuntimeFilters;
 pub use physical_limit::Limit;
 pub use physical_multi_table_insert::*;
 pub use physical_mutation::*;
